refactor(analysis): move `shift` to `transform`
The `shift` function has been moved from `utils` to `transform`. Calling from the `utils` module is now deprecated.
