+++ src/erlab/analysis/__init__.py
@@ -17,7 +17,6 @@ Modules
    interpolate
    kspace
    transform
-   utils
 
 """
