+++ src/erlab/analysis/gold.py
@@ -36,7 +36,7 @@ from erlab.analysis.fit.models import (
     PolynomialModel,
     StepEdgeModel,
 )
-from erlab.analysis.utils import shift
+from erlab.analysis.transform import shift
 from erlab.plotting.colors import proportional_colorbar
 from erlab.plotting.general import autoscale_to, figwh, plot_array
 from erlab.utils.parallel import joblib_progress
@@ -72,7 +72,7 @@ def correct_with_edge(
     plot_kw
         Additional keyword arguments for the plot. Defaults to `None`.
     **shift_kwargs
-        Additional keyword arguments to `erlab.analysis.utils.shift`.
+        Additional keyword arguments to :func:`erlab.analysis.transform.shift`.
 
     Returns
     -------
