--- tests/analysis/test_utils.py
@@ -1,25 +1,0 @@
-import numpy as np
-import xarray as xr
-from erlab.analysis.utils import shift
-
-
-def test_shift():
-    # Create a test input DataArray
-    darr = xr.DataArray(
-        np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]]).astype(float), dims=["x", "y"]
-    )
-
-    # Create a test shift DataArray
-    shift_arr = xr.DataArray([1, 0, 2], dims=["x"])
-
-    # Perform the shift operation
-    shifted = shift(darr, shift_arr, along="y")
-
-    # Define the expected result
-    expected = xr.DataArray(
-        np.array([[np.nan, 1.0, 2.0], [4.0, 5.0, 6.0], [np.nan, np.nan, 7.0]]),
-        dims=["x", "y"],
-    )
-
-    # Check if the shifted array matches the expected result
-    assert np.allclose(shifted, expected, equal_nan=True)
